// apps/fe-mobile/src/app/assets/images/type.ts
import { ImageRequireSource } from 'react-native';

export type ImagesTypes = {
  //APP LOGO
  stc_logo: ImageRequireSource;

  //BACKGROUND IMAGES
  bg_mask_purple: ImageRequireSource;

  //BOTTOM TAB
  discover: ImageRequireSource;
  compete: ImageRequireSource;
  connect: ImageRequireSource;
  shop: ImageRequireSource;
  add: ImageRequireSource;

  //ONBOARDING IMAGES
  follow_cash_bg: ImageRequireSource;
  follow_man_main: ImageRequireSource;
  score_controller_bg: ImageRequireSource;
  score_console_main: ImageRequireSource;
  connect_mobile_bg: ImageRequireSource;
  connect_man_main: ImageRequireSource;
  nevermiss_rectangle_bg: ImageRequireSource;
  nevermiss_man_main: ImageRequireSource;

  //RANDOM IMAGES
  profile_placeholder?: ImageRequireSource;
  display_photo?: ImageRequireSource;

  //ICON LIST
  icon_arrow_right: ImageRequireSource;
  icon_arrow_left: ImageRequireSource;
  icon_arrow: ImageRequireSource;
  twitter: ImageRequireSource;
  google: ImageRequireSource;
  facebook: ImageRequireSource;
  apple: ImageRequireSource;
  icon_check_circle: ImageRequireSource;
  icon_visibility: ImageRequireSource;
  icon_close: ImageRequireSource;
  icon_info: ImageRequireSource;
  img_Success: ImageRequireSource;
  icon_bell: ImageRequireSource;
  icon_cart: ImageRequireSource;
  icon_heart: ImageRequireSource;
  icon_magnifying: ImageRequireSource;
  icon_group: ImageRequireSource;
  icon_play_outline: ImageRequireSource;
};